allow in-place appending to .xlsx files
the natural workflow to convert a bunch of .xliff files into one
master.xlsx is to create the master.xlsx by the first use of `to-xlsx`
and then subsequentially `-append` to the same master.xlsx over and
over again.

to support this workflow, *xliffer* needs to read the contents of the
master.xlsx before it starts the convertion. before xliffer starts the
.Conver() function of every converter, it (optionally) opens a given
output file. this would truncate master.xlsx before the data inside
master.xlsx is read at all.

this commit changes the converter interface by adding a .Prepare()
function. the .Prepare() function is called before the convertion takes
place.

as a result, the user can now append to the same file over and over
again.