support <mrk> tags in <target> nodes
some translation services split <source> content into segments.
the segments are marked by <mrk> tags. *xliffer* before this
commit used only the chardata of the <target> node for all the
operations. chardata of these extra <mrk> nodes were ignored.

this commit changes the parsing of a .xliff file to extract the
chardata of <mrk> nodes below a <target> node and combine it into
the chardata of the <target> node itself:

    <source>
        This is Sparta! It really is
    </source>

This becomes something like this after some translation tools
are done with it:

    <target>
        <mrk mtype="seg">This is Sparta!</mrk> <mrk mtype="seg">It really is</mrk>
    </target>

This commit changes the parsing to yield this:

    <target>
        This is Sparta! It really is
    </target>