to-xlsx: improve robustness against erroneously user input
this commit affects the 'to-xlsx' converter of *xliff*:

a user wants to append .xliff translations to a given .xlsx file. the
translations should be integrated in a certain sheet which is given by
the -sheet flag.

this commit makes *xliffer* exit early if the given sheet name can not
be found in the .xlsx file. otherwise, the translations would end up on
the first sheet which might mess up existing translations.